keylayout: swap volume buttons
due to the default screenrotation of 270 degrees the volume buttons are not physically aligend to the graphical volume slider. fix this by swapping vol+ and vol-.
for the correct orientation of the volume buttons in portrait one can check the "swap volume buttons when the screen is rotated"-switch in settings>display>rotation.